pyproject.toml
[tool.poetry]
name = "notes-api"
version = "0.1.0"
description = ""
authors = ["poligrafo <poligrafo888@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.38"
asyncpg = "^0.30.0"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
alembic = "^1.14.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
pyjwt = "^2.10.1"
loguru = "^0.7.3"
pytest = "^8.3.4"
httpx = "^0.28.1"
starlette-admin = "^0.14.1"
sqlalchemy-admin = "^0.1.2"
itsdangerous = "^2.2.0"
sqladmin = "^0.20.1"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
